fixed comments from  weekly review meeting
