+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon --exec ts-node src/Week3/firebaseUser.ts",
+    "start": "nodemon --exec ts-node src/Week4/index.ts",
     "test": "jest"
   },
   "keywords": [],
